6Back/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
numpy==1.26.4
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tensorflow==2.17.0
whitenoise==6.11.0
